+++ CHANGELOG.md
@@ -37,13 +37,15 @@ Ref: https://keepachangelog.com/en/1.0.0/
 
 ## [Unreleased]
 
+### Bug Fixes
+
+* (modules/service) [\#18](https://github.com/irisnet/irismod/pull/18) Fix incorrect price in request and fix init request when insufficient banlances.
+
 ## [v1.1.0] - 2020-09-30
 
 ### Features
 
 * Add modules `token`, `nft`, `htlc`, `coinswap`, `service`, `oracle`, `random`, `record`.
-
-
 
 <!-- Release links -->